cache size use int64

+++ cache.go
@@ -9,7 +9,7 @@ type Cache struct {
 }
 
 // NewLRUCache creates a new LRU Cache object with the capacity given.
-func NewLRUCache(capacity int) *Cache {
+func NewLRUCache(capacity int64) *Cache {
 	return NewNativeCache(C.rocksdb_cache_create_lru(C.size_t(capacity)))
 }
